third_party/WebKit/LayoutTests/fast/css/invalidation/content-pseudo-expected.txt

CONSOLE WARNING: line 11: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
No need to invalidate for selectors right of ::content as ::content causes subtree invalidation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(green).color is "rgb(0, 0, 0)"
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(green).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

This text should be green
